# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Tezos Quest Indexer",
        "description": "API endpoints for Quest Indexer"
    },
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/rewards_pending/{game_id}": {
            "get": {
                "summary": "gets the all the pending FA2 transfers needed to be sent by an admin to the FA2 contract",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "game_id",
                        "in": "query",
                        "description": "the game_id to check for any pending reward",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "this overrides the tezos FA2 contract account you would to send FA2 from",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "---": {
                                        "status": "success",
                                        "data": {
                                            "humanArgs": [
                                                [
                                                    {
                                                        "from_": "tz1NXuEyKGjGXAEn5gSqzgwhkEqNeN5wcaLf",
                                                        "txs": [
                                                            {
                                                                "to_": "tz1Myo4eJ6JaGEPkTJw3cf1R3MQEmZQmkGt4",
                                                                "token_id": 20,
                                                                "amount": 0
                                                            },
                                                            {
                                                                "to_": "tz1PirboHQVqkYqLSWfHUHEy3AdhYUNJpvGy",
                                                                "token_id": 13,
                                                                "amount": 0
                                                            }
                                                        ]
                                                    }
                                                ]
                                            ],
                                            "transferParamsFA2": [
                                                {
                                                    "to": "KT1PS2jZVzNMW54UsnqBqwwkArXnAZ29jiTF",
                                                    "amount": 0,
                                                    "mutez": false,
                                                    "parameter": {
                                                        "entrypoint": "transfer",
                                                        "value": [
                                                            {
                                                                "prim": "Pair",
                                                                "args": [
                                                                    {
                                                                        "string": "tz1NXuEyKGjGXAEn5gSqzgwhkEqNeN5wcaLf"
                                                                    },
                                                                    [
                                                                        {
                                                                            "prim": "Pair",
                                                                            "args": [
                                                                                {
                                                                                    "string": "tz1Myo4eJ6JaGEPkTJw3cf1R3MQEmZQmkGt4"
                                                                                },
                                                                                {
                                                                                    "prim": "Pair",
                                                                                    "args": [
                                                                                        {
                                                                                            "int": "20"
                                                                                        },
                                                                                        {
                                                                                            "int": "0"
                                                                                        }
                                                                                    ]
                                                                                }
                                                                            ]
                                                                        },
                                                                        {
                                                                            "prim": "Pair",
                                                                            "args": [
                                                                                {
                                                                                    "string": "tz1PirboHQVqkYqLSWfHUHEy3AdhYUNJpvGy"
                                                                                },
                                                                                {
                                                                                    "prim": "Pair",
                                                                                    "args": [
                                                                                        {
                                                                                            "int": "13"
                                                                                        },
                                                                                        {
                                                                                            "int": "0"
                                                                                        }
                                                                                    ]
                                                                                }
                                                                            ]
                                                                        }
                                                                    ]
                                                                ]
                                                            }
                                                        ]
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/set_filter_dates": {
            "post": {
                "summary": "updates dates in operation filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "filter_id": {
                                        "type": "number",
                                        "example": "12"
                                    },
                                    "time_start": {
                                        "type": "string",
                                        "example": "2021-03-26T23:07:04.405Z"
                                    },
                                    "time_end": {
                                        "type": "string",
                                        "example": "2021-03-26T23:07:04.405Z"
                                    }
                                },
                                "required": [
                                    "filter_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}